+++ stroom-core-client/src/main/java/stroom/dispatch/client/ApplicationInstanceIdProvider.java
@@ -1,6 +1,6 @@
 package stroom.dispatch.client;
 
-import stroom.util.client.RandomId;
+import stroom.util.shared.RandomId;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;